fix(ci): restore abort cleanup checks
