+++ ci/source-architecture-budget.json
@@ -8,14 +8,14 @@
       "src/lib/adapters/docker/index.ts": 43,
       "src/lib/adapters/openshell/client.ts": 23,
       "src/lib/adapters/openshell/resolve.ts": 27,
-      "src/lib/adapters/openshell/runtime.ts": 52,
+      "src/lib/adapters/openshell/runtime.ts": 53,
       "src/lib/adapters/openshell/timeouts.ts": 37,
       "src/lib/agent/defs.ts": 32,
       "src/lib/cli/branding.ts": 86,
       "src/lib/cli/nemoclaw-oclif-command.ts": 106,
       "src/lib/cli/terminal-style.ts": 43,
       "src/lib/core/json-types.ts": 37,
-      "src/lib/core/ports.ts": 88,
+      "src/lib/core/ports.ts": 89,
       "src/lib/core/shell-quote.ts": 28,
       "src/lib/core/url-utils.ts": 27,
       "src/lib/core/wait.ts": 35,
@@ -23,11 +23,11 @@
       "src/lib/inference/config.ts": 29,
       "src/lib/inference/web-search.ts": 21,
       "src/lib/messaging/channels/index.ts": 25,
-      "src/lib/onboard/gateway-binding.ts": 49,
+      "src/lib/onboard/gateway-binding.ts": 50,
       "src/lib/runner.ts": 88,
       "src/lib/security/redact.ts": 52,
       "src/lib/state/onboard-session.ts": 36,
-      "src/lib/state/registry.ts": 99,
+      "src/lib/state/registry.ts": 100,
       "src/lib/state/state-root.ts": 20,
       "src/lib/subprocess-env.ts": 24,
       "src/lib/validation.ts": 25
@@ -46,7 +46,7 @@
       "src/lib/actions/sandbox/rebuild-pipeline.ts": 28,
       "src/lib/actions/sandbox/snapshot.ts": 40,
       "src/lib/actions/uninstall/run-plan.ts": 26,
-      "src/lib/inference/onboard-probes.ts": 20,
+      "src/lib/inference/onboard-probes.ts": 21,
       "src/lib/inference/vllm.ts": 21,
       "src/lib/onboard.ts": 210,
       "src/lib/onboard/machine/handlers/sandbox.ts": 21,
